+++ src/main/kotlin/ecommerce/config/LoginMemberInterceptor.kt
@@ -1,0 +1,30 @@
+package ecommerce.config
+
+import ecommerce.enums.UserRole
+import ecommerce.exception.UnauthorisedUserException
+import ecommerce.infrastructure.JwtTokenProvider
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
+import org.springframework.stereotype.Component
+import org.springframework.web.servlet.HandlerInterceptor
+
+annotation class LoginMember
+
+@Component
+class LoginMemberInterceptor(
+    private val jwtTokenProvider: JwtTokenProvider,
+) : HandlerInterceptor {
+    override fun preHandle(
+        request: HttpServletRequest,
+        response: HttpServletResponse,
+        handler: Any,
+    ): Boolean {
+        val bearer = request.getHeader("Authorization") ?: throw UnauthorisedUserException()
+        val token = bearer.replace("Bearer ", "").trim()
+        val payload = jwtTokenProvider.getPayload(token)
+        if (payload.role == UserRole.USER) {
+            throw UnauthorisedUserException("Admin role required")
+        }
+        return true
+    }
+}
