+++ src/main/kotlin/ecommerce/config/WebConfig.kt
@@ -1,0 +1,16 @@
+package ecommerce.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+
+@Configuration
+class WebConfig(
+    private val LoginMemberInterceptor: LoginMemberInterceptor,
+) : WebMvcConfigurer {
+    override fun addInterceptors(registry: InterceptorRegistry) {
+//        registry.addInterceptor(LoginMemberInterceptor)
+//            .addPathPatterns("/api/products/**")=
+        super.addInterceptors(registry)
+    }
+}
