+++ src/main/kotlin/ecommerce/dto/auth/TokenRequest.kt
@@ -1,4 +1,4 @@
-package ecommerce.dto
+package ecommerce.dto.auth
 
 data class TokenRequest(
     val email: String,
