+++ src/main/kotlin/ecommerce/dto/auth/TokenResponse.kt
@@ -1,4 +1,4 @@
-package ecommerce.dto
+package ecommerce.dto.auth
 
 data class TokenResponse(
     val token: String,
