+++ src/main/kotlin/ecommerce/dto/error/ErrorResponse.kt
@@ -1,4 +1,4 @@
-package ecommerce.dto
+package ecommerce.dto.error
 
 import java.time.Instant
