+++ src/test/kotlin/ecommerce/repository/ProductRepositoryTest.kt
@@ -1,6 +1,6 @@
 package ecommerce.repository
 
-import ecommerce.dto.ProductDTO
+import ecommerce.dto.products.ProductDTO
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
